refactor(review): 粒度別ビューの dispatcher 構造に ReviewView を再編
review-granularity-redesign Step 1。既存サマリーレイアウトを
views/WeeklyReview.tsx へ移動し、ReviewView は granularity で
composition を切り替える dispatcher にする。day / month / year は
固有ビュー実装（Step 3-4）まで WeeklyReview を暫定表示（見た目の
regression なし）。
